refactor nodeclass controller
