+++ pkg/controllers/nodeclass/status/instanceprofile.go
@@ -1,0 +1,44 @@
+/*
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package status
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/samber/lo"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	"github.com/aws/karpenter-provider-aws/pkg/apis/v1beta1"
+	"github.com/aws/karpenter-provider-aws/pkg/providers/instanceprofile"
+)
+
+type InstanceProfile struct {
+	instanceProfileProvider instanceprofile.Provider
+}
+
+func (ip *InstanceProfile) Reconcile(ctx context.Context, nodeClass *v1beta1.EC2NodeClass) (reconcile.Result, error) {
+	if nodeClass.Spec.Role != "" {
+		name, err := ip.instanceProfileProvider.Create(ctx, nodeClass)
+		if err != nil {
+			return reconcile.Result{}, fmt.Errorf("creating instance profile, %w", err)
+		}
+		nodeClass.Status.InstanceProfile = name
+	} else {
+		nodeClass.Status.InstanceProfile = lo.FromPtr(nodeClass.Spec.InstanceProfile)
+	}
+
+	return reconcile.Result{}, nil
+}
