+++ pkg/controllers/nodeclass/termination/suite_test.go
@@ -1,0 +1,285 @@
+/*
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package termination_test
+
+import (
+	"context"
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/samber/lo"
+	"k8s.io/client-go/tools/record"
+	_ "knative.dev/pkg/system/testing"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/ec2"
+	"github.com/aws/aws-sdk-go/service/iam"
+	corev1beta1 "sigs.k8s.io/karpenter/pkg/apis/v1beta1"
+	"sigs.k8s.io/karpenter/pkg/events"
+	corecontroller "sigs.k8s.io/karpenter/pkg/operator/controller"
+	coreoptions "sigs.k8s.io/karpenter/pkg/operator/options"
+	"sigs.k8s.io/karpenter/pkg/operator/scheme"
+	coretest "sigs.k8s.io/karpenter/pkg/test"
+
+	"github.com/aws/karpenter-provider-aws/pkg/apis"
+	"github.com/aws/karpenter-provider-aws/pkg/apis/v1beta1"
+	"github.com/aws/karpenter-provider-aws/pkg/controllers/nodeclass/termination"
+	"github.com/aws/karpenter-provider-aws/pkg/fake"
+	"github.com/aws/karpenter-provider-aws/pkg/operator/options"
+	"github.com/aws/karpenter-provider-aws/pkg/test"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	. "knative.dev/pkg/logging/testing"
+	. "sigs.k8s.io/karpenter/pkg/test/expectations"
+)
+
+var ctx context.Context
+var env *coretest.Environment
+var awsEnv *test.Environment
+var terminationController corecontroller.Controller
+
+func TestAPIs(t *testing.T) {
+	ctx = TestContextWithLogger(t)
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "EC2NodeClass")
+}
+
+var _ = BeforeSuite(func() {
+	env = coretest.NewEnvironment(scheme.Scheme, coretest.WithCRDs(apis.CRDs...), coretest.WithFieldIndexers(test.EC2NodeClassFieldIndexer(ctx)))
+	ctx = coreoptions.ToContext(ctx, coretest.Options())
+	ctx = options.ToContext(ctx, test.Options())
+	awsEnv = test.NewEnvironment(ctx, env)
+
+	terminationController = termination.NewController(env.Client, events.NewRecorder(&record.FakeRecorder{}), awsEnv.InstanceProfileProvider, awsEnv.LaunchTemplateProvider)
+})
+
+var _ = AfterSuite(func() {
+	Expect(env.Stop()).To(Succeed(), "Failed to stop environment")
+})
+
+var _ = BeforeEach(func() {
+	ctx = coreoptions.ToContext(ctx, coretest.Options())
+	awsEnv.Reset()
+})
+
+var _ = AfterEach(func() {
+	ExpectCleanedUp(ctx, env.Client)
+})
+
+var _ = Describe("NodeClass Termination", func() {
+	var nodeClass *v1beta1.EC2NodeClass
+	var profileName string
+	BeforeEach(func() {
+		nodeClass = test.EC2NodeClass(v1beta1.EC2NodeClass{
+			Spec: v1beta1.EC2NodeClassSpec{
+				SubnetSelectorTerms: []v1beta1.SubnetSelectorTerm{
+					{
+						Tags: map[string]string{"*": "*"},
+					},
+				},
+				SecurityGroupSelectorTerms: []v1beta1.SecurityGroupSelectorTerm{
+					{
+						Tags: map[string]string{"*": "*"},
+					},
+				},
+				AMISelectorTerms: []v1beta1.AMISelectorTerm{
+					{
+						Tags: map[string]string{"*": "*"},
+					},
+				},
+			},
+		})
+		profileName = awsEnv.InstanceProfileProvider.GetProfileName(ctx, fake.DefaultRegion, nodeClass.Name)
+	})
+	It("should not delete the NodeClass if launch template deletion fails", func() {
+		launchTemplateName := aws.String(fake.LaunchTemplateName())
+		awsEnv.EC2API.LaunchTemplates.Store(launchTemplateName, &ec2.LaunchTemplate{LaunchTemplateName: launchTemplateName, LaunchTemplateId: aws.String(fake.LaunchTemplateID()), Tags: []*ec2.Tag{&ec2.Tag{Key: aws.String("karpenter.k8s.aws/cluster"), Value: aws.String("test-cluster")}}})
+		_, ok := awsEnv.EC2API.LaunchTemplates.Load(launchTemplateName)
+		Expect(ok).To(BeTrue())
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		awsEnv.EC2API.NextError.Set(fmt.Errorf("delete Launch Template Error"))
+		ExpectReconcileFailed(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		ExpectExists(ctx, env.Client, nodeClass)
+	})
+	It("should not delete the launch template not associated with the nodeClass", func() {
+		launchTemplateName := aws.String(fake.LaunchTemplateName())
+		awsEnv.EC2API.LaunchTemplates.Store(launchTemplateName, &ec2.LaunchTemplate{LaunchTemplateName: launchTemplateName, LaunchTemplateId: aws.String(fake.LaunchTemplateID()), Tags: []*ec2.Tag{&ec2.Tag{Key: aws.String("karpenter.k8s.aws/cluster"), Value: aws.String("test-cluster")}}})
+		_, ok := awsEnv.EC2API.LaunchTemplates.Load(launchTemplateName)
+		Expect(ok).To(BeTrue())
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		_, ok = awsEnv.EC2API.LaunchTemplates.Load(launchTemplateName)
+		Expect(ok).To(BeTrue())
+		ExpectNotFound(ctx, env.Client, nodeClass)
+	})
+	It("should succeed to delete the launch template", func() {
+		ltName1 := aws.String(fake.LaunchTemplateName())
+		awsEnv.EC2API.LaunchTemplates.Store(ltName1, &ec2.LaunchTemplate{LaunchTemplateName: ltName1, LaunchTemplateId: aws.String(fake.LaunchTemplateID()), Tags: []*ec2.Tag{&ec2.Tag{Key: aws.String("karpenter.k8s.aws/cluster"), Value: aws.String("test-cluster")}, {Key: aws.String("karpenter.k8s.aws/ec2nodeclass"), Value: aws.String(nodeClass.Name)}}})
+		ltName2 := aws.String(fake.LaunchTemplateName())
+		awsEnv.EC2API.LaunchTemplates.Store(ltName2, &ec2.LaunchTemplate{LaunchTemplateName: ltName2, LaunchTemplateId: aws.String(fake.LaunchTemplateID()), Tags: []*ec2.Tag{&ec2.Tag{Key: aws.String("karpenter.k8s.aws/cluster"), Value: aws.String("test-cluster")}, {Key: aws.String("karpenter.k8s.aws/ec2nodeclass"), Value: aws.String(nodeClass.Name)}}})
+		_, ok := awsEnv.EC2API.LaunchTemplates.Load(ltName1)
+		Expect(ok).To(BeTrue())
+		_, ok = awsEnv.EC2API.LaunchTemplates.Load(ltName2)
+		Expect(ok).To(BeTrue())
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		_, ok = awsEnv.EC2API.LaunchTemplates.Load(ltName1)
+		Expect(ok).To(BeFalse())
+		_, ok = awsEnv.EC2API.LaunchTemplates.Load(ltName2)
+		Expect(ok).To(BeFalse())
+		ExpectNotFound(ctx, env.Client, nodeClass)
+	})
+	It("should succeed to delete the instance profile with no NodeClaims", func() {
+		awsEnv.IAMAPI.InstanceProfiles = map[string]*iam.InstanceProfile{
+			profileName: {
+				InstanceProfileName: aws.String(profileName),
+				Roles: []*iam.Role{
+					{
+						RoleId:   aws.String(fake.RoleID()),
+						RoleName: aws.String(nodeClass.Spec.Role),
+					},
+				},
+			},
+		}
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(0))
+		ExpectNotFound(ctx, env.Client, nodeClass)
+	})
+	It("should succeed to delete the instance profile when no roles exist with no NodeClaims", func() {
+		awsEnv.IAMAPI.InstanceProfiles = map[string]*iam.InstanceProfile{
+			profileName: {
+				InstanceProfileName: aws.String(profileName),
+			},
+		}
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(0))
+		ExpectNotFound(ctx, env.Client, nodeClass)
+	})
+	It("should succeed to delete the NodeClass when the instance profile doesn't exist", func() {
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(0))
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(0))
+		ExpectNotFound(ctx, env.Client, nodeClass)
+	})
+	It("should not delete the EC2NodeClass until all associated NodeClaims are terminated", func() {
+		var nodeClaims []*corev1beta1.NodeClaim
+		for i := 0; i < 2; i++ {
+			nc := coretest.NodeClaim(corev1beta1.NodeClaim{
+				Spec: corev1beta1.NodeClaimSpec{
+					NodeClassRef: &corev1beta1.NodeClassReference{
+						Name: nodeClass.Name,
+					},
+				},
+			})
+			ExpectApplied(ctx, env.Client, nc)
+			nodeClaims = append(nodeClaims, nc)
+		}
+		awsEnv.IAMAPI.InstanceProfiles = map[string]*iam.InstanceProfile{
+			profileName: {
+				InstanceProfileName: aws.String(profileName),
+				Roles: []*iam.Role{
+					{
+						RoleId:   aws.String(fake.RoleID()),
+						RoleName: aws.String(nodeClass.Spec.Role),
+					},
+				},
+			},
+		}
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		res := ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(res.RequeueAfter).To(Equal(time.Minute * 10))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+		ExpectExists(ctx, env.Client, nodeClass)
+
+		// Delete one of the NodeClaims
+		// The NodeClass should still not delete
+		ExpectDeleted(ctx, env.Client, nodeClaims[0])
+		res = ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(res.RequeueAfter).To(Equal(time.Minute * 10))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+		ExpectExists(ctx, env.Client, nodeClass)
+
+		// Delete the last NodeClaim
+		// The NodeClass should now delete
+		ExpectDeleted(ctx, env.Client, nodeClaims[1])
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(0))
+		ExpectNotFound(ctx, env.Client, nodeClass)
+	})
+	It("should not call the IAM API when deleting a NodeClass with an instanceProfile specified", func() {
+		awsEnv.IAMAPI.InstanceProfiles = map[string]*iam.InstanceProfile{
+			profileName: {
+				InstanceProfileName: aws.String("test-instance-profile"),
+				Roles: []*iam.Role{
+					{
+						RoleId:   aws.String(fake.RoleID()),
+						RoleName: aws.String("fake-role"),
+					},
+				},
+			},
+		}
+		nodeClass.Spec.Role = ""
+		nodeClass.Spec.InstanceProfile = lo.ToPtr("test-instance-profile")
+		controllerutil.AddFinalizer(nodeClass, v1beta1.TerminationFinalizer)
+		ExpectApplied(ctx, env.Client, nodeClass)
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+
+		Expect(env.Client.Delete(ctx, nodeClass)).To(Succeed())
+		ExpectReconcileSucceeded(ctx, terminationController, client.ObjectKeyFromObject(nodeClass))
+		Expect(awsEnv.IAMAPI.InstanceProfiles).To(HaveLen(1))
+		ExpectNotFound(ctx, env.Client, nodeClass)
+
+		Expect(awsEnv.IAMAPI.DeleteInstanceProfileBehavior.Calls()).To(BeZero())
+		Expect(awsEnv.IAMAPI.RemoveRoleFromInstanceProfileBehavior.Calls()).To(BeZero())
+	})
+})
